Move dbconsole under namespace db
